+++ Scripts/run.js
@@ -137,7 +137,7 @@ function isCombinedTestRun(params) {
 const isCombinedTest = isCombinedTestRun(params);
 process.env.CYPRESS_isCombinedTestRun = isCombinedTest;
 
-// Extract jobId from the parameters
+// Extract jobId and logger level from the parameters
 let jobId = '';
 const processingEnvArgs = params.includes('--env');
 if (processingEnvArgs) {
@@ -145,7 +145,10 @@ if (processingEnvArgs) {
   for (const envArg of envArgs) {
     if (envArg.startsWith('jobId=')) {
       jobId = envArg.split('=')[1];
-      break;
+    }
+    if (envArg.startsWith('loggerLevel=')) {
+      const loggerLevelFromEnv = envArg.split('=')[1];
+      process.env.loggerLevel = loggerLevelFromEnv;
     }
   }
 }